Initial Setup
Just trying to make this a folder with a gitignore file and a Unity project. Hopefully said project is openable--it had issues that it resolved first time but other issues like the scenes just plain not showing UI elements are yet to be seen fixed, until I reopen the project.