Whoops, we've used a lot of time seemingly making buttons small and text boxes tall.
Actually have a more solid understanding of the story, the characters, and got pretty much to the level of functionality we lost last jam. Because I'm real bad at remembering how my code requires prefabs nest and not nest.

Next to transition to actually writing this story so it can fit with the sound concepts I need to test.