Volume control
Once I get the sounds into the game, they have volume sliders to control them and I just have to add them all to a list and check whether they're SFX or not. Certainly more work to do in getting sound files to play (and stop playing) in relation to portions of story, but that was a good three hours of failing to cobble together two tutorials for a result I'm going to find useful. Here and in future projects.

I knew my brain was slow these past few weeks, but my goodness. Still. Things are happening.